+++ src/core/algorithm/diskann/diskann_index.cc
@@ -17,6 +17,7 @@
 #include <iostream>
 #include <memory>
 #include <set>
+#include <algorithm>
 
 namespace zvec {
 namespace core {
@@ -315,7 +316,7 @@ void DiskAnnIndex::cache_bfs_levels(uint64_t num_nodes_to_cache,
         DiskAnnUtil::div_round_up(nodes_to_expand.size(), BLOCK_SIZE);
     for (size_t block = 0; block < nblocks && !finish_flag; block++) {
       size_t start = block * BLOCK_SIZE;
-      size_t end = std::min((block + 1) * BLOCK_SIZE, nodes_to_expand.size());
+      size_t end = std::min((uint64_t)((block + 1) * BLOCK_SIZE), (uint64_t)(nodes_to_expand.size()));
 
       std::vector<diskann_id_t> nodes_to_read;
       std::vector<void *> coord_buffers(end - start, nullptr);
